Log response status code and duration
