+++ server/middleware/context_test.go
@@ -8,8 +8,6 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"github.com/root-gg/plik/server/common"
-
 	"github.com/root-gg/plik/server/context"
 )
 
@@ -30,38 +28,4 @@ func TestContext(t *testing.T) {
 
 	context.NewChain(Context(setup)).Then(handler).ServeHTTP(rr, req)
 	require.True(t, ok, "handler was not called")
-}
-
-func TestPanic(t *testing.T) {
-
-	config := &common.Configuration{}
-	setup := func(ctx *context.Context) { ctx.SetConfig(config) }
-
-	handler := func(ctx *context.Context, resp http.ResponseWriter, req *http.Request) {
-		panic("defuk")
-	}
-
-	rr := httptest.NewRecorder()
-	req, err := http.NewRequest("GET", "", &bytes.Buffer{})
-	require.NoError(t, err, "unable to create new request")
-
-	context.NewChain(Context(setup)).Then(handler).ServeHTTP(rr, req)
-	context.TestInternalServerError(t, rr, "internal server error")
-}
-
-func TestPanicDebug(t *testing.T) {
-
-	config := &common.Configuration{Debug: true}
-	setup := func(ctx *context.Context) { ctx.SetConfig(config) }
-
-	handler := func(ctx *context.Context, resp http.ResponseWriter, req *http.Request) {
-		panic("defuk")
-	}
-
-	rr := httptest.NewRecorder()
-	req, err := http.NewRequest("GET", "", &bytes.Buffer{})
-	require.NoError(t, err, "unable to create new request")
-
-	context.NewChain(Context(setup)).Then(handler).ServeHTTP(rr, req)
-	context.TestInternalServerError(t, rr, "panic : defuk")
 }
